Add context with recent files
